+++ app/Http/Controllers/Examples/SeniorityEnum.php
@@ -11,7 +11,7 @@ class SeniorityEnum extends Enum
     protected static $data = [
         1 => 'Assistant',
         2 => 'Associate',
-        3 => 'Staff Member',
+        3 => 'Staff',
         4 => 'Senior',
         5 => 'Partner',
     ];