+++ config/enso/config.php
@@ -1,7 +1,7 @@
 <?php
 
 return [
-    'version' => '2.3.9',
+    'version' => '2.3.10',
     'defaultRole' => 'admin',
     'phpDateFormat' => 'd-m-Y',
     'jsDateFormat' => 'DD-MM-YYYY',