+++ resources/views/examples/partials/vueSelectFilter.blade.php
@@ -3,6 +3,7 @@
     <code class="js">
 vue-select-filter class="box"
     title="Seniority"
+    multiple
     :options="seniorityOptions"
     v-model="filters.examples.seniority"
 vue-select-filter