+++ resources/views/examples/table.blade.php
@@ -50,7 +50,7 @@
             </section>
             <div class="container has-margin-top-large">
                 <div class="columns">
-                    <div class="column is-one-fifth">
+                    <div class="column is-2">
                         <vue-filter class="box"
                             :options="activeOptions"
                             icons
@@ -58,14 +58,15 @@
                             v-model="filters.examples.is_active">
                         </vue-filter>
                     </div>
-                    <div class="column is-one-fifth">
+                    <div class="column is-3">
                         <vue-select-filter class="box"
                             title="Seniority"
+                            multiple
                             :options="seniorityOptions"
                             v-model="filters.examples.seniority">
                         </vue-select-filter>
                     </div>
-                    <div class="column is-two-fifths">
+                    <div class="column is-4">
                         <date-interval-filter class="box"
                             title="Hired Between"
                             :min="intervals.examples.hired_at.min"
@@ -74,7 +75,7 @@
                             @update-max="intervals.examples.hired_at.max = $event">
                         </date-interval-filter>
                     </div>
-                    <div class="column is-one-fifth">
+                    <div class="column is-3">
                         <interval-filter class="box"
                             title="Salary"
                             type="number"